# internal/infrastructure/http/server/server.go
package server

import (
	"context"
	"database/sql"
	"fmt"
	"net/http"
	"strings"
	"time"

	"github.com/go-chi/chi/v5"
	"github.com/go-chi/cors"
	"github.com/matt-riley/mjrwtf/internal/adapters/notification"
	"github.com/matt-riley/mjrwtf/internal/adapters/repository"
	"github.com/matt-riley/mjrwtf/internal/application"
	"github.com/matt-riley/mjrwtf/internal/domain/click"
	"github.com/matt-riley/mjrwtf/internal/domain/url"
	"github.com/matt-riley/mjrwtf/internal/domain/urlstatus"
	"github.com/matt-riley/mjrwtf/internal/infrastructure/config"
	"github.com/matt-riley/mjrwtf/internal/infrastructure/http/handlers"
	"github.com/matt-riley/mjrwtf/internal/infrastructure/http/middleware"
	"github.com/matt-riley/mjrwtf/internal/infrastructure/metrics"
	"github.com/matt-riley/mjrwtf/internal/infrastructure/session"
	"github.com/rs/zerolog"
)

const (
	// Server timeout configurations
	readTimeout  = 15 * time.Second
	writeTimeout = 15 * time.Second
	idleTimeout  = 60 * time.Second

	// ShutdownTimeout is the maximum time allowed for graceful server shutdown.
	ShutdownTimeout = 30 * time.Second

	defaultRedirectRateLimitPerMinute = 120
	defaultAPIRateLimitPerMinute      = 60
)

// Server represents the HTTP server
type Server struct {
	httpServer       *http.Server
	router           *chi.Mux
	config           *config.Config
	db               *sql.DB
	logger           zerolog.Logger
	metrics          *metrics.Metrics
	sessionStore     *session.Store
	rateLimiters     []*middleware.RateLimiterMiddleware
	redirectUseCase  *application.RedirectURLUseCase
	urlStatusChecker *application.URLStatusChecker

	bgCtx    context.Context
	bgCancel context.CancelFunc
}

// New creates a new HTTP server with configured middleware and dependencies
// Returns an error if the server cannot be initialized properly
func New(cfg *config.Config, db *sql.DB, logger zerolog.Logger) (*Server, error) {
	r := chi.NewRouter()

	// Initialize Prometheus metrics
	m := metrics.New()

	// Initialize Discord notifier for error notifications
	var discordNotifier *notification.DiscordNotifier
	if cfg.DiscordWebhookURL != "" {
		discordNotifier = notification.NewDiscordNotifier(
			cfg.DiscordWebhookURL,
			notification.WithLogger(logger),
		)
		logger.Info().Msg("Discord error notifications enabled")
	} else {
		logger.Info().Msg("Discord error notifications disabled (no webhook URL configured)")
	}

	// Middleware stack (order matters)
	r.Use(middleware.RecoveryWithNotifier(logger, discordNotifier)) // Recover from panics first, with Discord notifications
	r.Use(middleware.RequestID)                                     // Generate/propagate request ID
	r.Use(middleware.SecurityHeaders(cfg.EnableHSTS))               // Set security headers
	r.Use(middleware.InjectLogger(logger))                          // Inject logger with request context
	r.Use(middleware.Logger)                                        // Log all requests
	r.Use(middleware.PrometheusMetrics(m))                          // Record Prometheus metrics

	// Initialize session store (24 hour session TTL)
	sessionStore := session.NewStore(24 * time.Hour)

	// Add session middleware globally (checks for session, but doesn't require it)
	r.Use(middleware.SessionMiddleware(sessionStore))

	// Parse CORS allowed origins (supports comma-separated list)
	origins := strings.Split(cfg.AllowedOrigins, ",")
	for i := range origins {
		origins[i] = strings.TrimSpace(origins[i])
	}

	r.Use(cors.Handler(cors.Options{
		// Note: Using "*" for allowed origins is a security risk in production.
		// Configure ALLOWED_ORIGINS environment variable to restrict access to known domains.
		AllowedOrigins:   origins,
		AllowedMethods:   []string{"GET", "POST", "PUT", "DELETE", "OPTIONS"},
		AllowedHeaders:   []string{"Accept", "Authorization", "Content-Type", "X-Request-ID"},
		ExposedHeaders:   []string{"Link", "X-Request-ID"},
		AllowCredentials: false,
		MaxAge:           300,
	}))

	bgCtx, bgCancel := context.WithCancel(context.Background())

	server := &Server{
		router:       r,
		config:       cfg,
		db:           db,
		logger:       logger,
		metrics:      m,
		sessionStore: sessionStore,
		bgCtx:        bgCtx,
		bgCancel:     bgCancel,
		httpServer: &http.Server{
			Addr:         fmt.Sprintf(":%d", cfg.ServerPort),
			Handler:      r,
			ReadTimeout:  readTimeout,
			WriteTimeout: writeTimeout,
			IdleTimeout:  idleTimeout,
		},
	}

	// Setup routes
	if err := server.setupRoutes(); err != nil {
		return nil, err
	}

	return server, nil
}

// setupRoutes configures the HTTP routes
func (s *Server) setupRoutes() error {
	redirectRateLimiter, apiRateLimiter := s.setupRateLimiters()

	s.setupHealthRoutes()
	s.setupMetricsRoutes()

	h, err := s.buildHandlers()
	if err != nil {
		return err
	}

	s.setupPageRoutes(h.pageHandler)

	// Public redirect endpoint (no authentication required)
	// Must come after specific routes to avoid capturing them
	s.setupRedirectRoutes(h.redirectHandler, redirectRateLimiter)

	s.setupAPIRoutes(h.urlHandler, h.analyticsHandler, apiRateLimiter)

	return nil
}

type routeHandlers struct {
	urlHandler       *handlers.URLHandler
	analyticsHandler *handlers.AnalyticsHandler
	redirectHandler  *handlers.RedirectHandler
	pageHandler      *handlers.PageHandler
}

func (s *Server) setupRateLimiters() (*middleware.RateLimiterMiddleware, *middleware.RateLimiterMiddleware) {
	// Defensive defaults: server.New can be called with a manually-constructed config
	// (e.g. in tests), which bypasses config.LoadConfig() validation.
	redirectRateLimit := s.config.RedirectRateLimitPerMinute
	if redirectRateLimit <= 0 {
		redirectRateLimit = defaultRedirectRateLimitPerMinute
	}

	apiRateLimit := s.config.APIRateLimitPerMinute
	if apiRateLimit <= 0 {
		apiRateLimit = defaultAPIRateLimitPerMinute
	}

	redirectRateLimiter := middleware.NewRateLimiterMiddleware(redirectRateLimit, time.Minute)
	apiRateLimiter := middleware.NewRateLimiterMiddleware(apiRateLimit, time.Minute)

	s.rateLimiters = []*middleware.RateLimiterMiddleware{redirectRateLimiter, apiRateLimiter}

	return redirectRateLimiter, apiRateLimiter
}

func (s *Server) setupHealthRoutes() {
	// Health check endpoint (liveness)
	// This is a lightweight check that does not validate external dependencies.
	s.router.Get("/health", s.healthCheckHandler)

	// Readiness endpoint
	// This validates required dependencies (e.g. database connectivity).
	s.router.Get("/ready", s.readyCheckHandler)
}

func (s *Server) setupMetricsRoutes() {
	// Prometheus metrics endpoint
	// Note: Authentication can be enabled via METRICS_AUTH_ENABLED environment variable.
	// In production, either enable authentication or restrict access via network policies/reverse proxy.
	// The endpoint exposes operational metrics (request rates, error rates, etc.)
	// which may be sensitive. Apply authentication if exposed to the public internet.
	if s.config.MetricsAuthEnabled {
		s.router.With(middleware.Auth(s.config.ActiveAuthTokens())).Handle("/metrics", s.metrics.Handler())
		return
	}

	s.router.Handle("/metrics", s.metrics.Handler())
}

func (s *Server) buildHandlers() (*routeHandlers, error) {
	// Initialize repositories (SQLite only)
	var urlRepo url.Repository = repository.NewSQLiteURLRepository(s.db)
	var clickRepo click.Repository = repository.NewSQLiteClickRepository(s.db)
	var urlStatusRepo urlstatus.Repository = repository.NewSQLiteURLStatusRepository(s.db)

	// Defensive defaults: server.New can be called with a manually-constructed config
	// (e.g. in tests), which bypasses config.LoadConfig() validation/defaults.
	dbTimeout := s.config.DBTimeout
	if dbTimeout <= 0 {
		dbTimeout = 5 * time.Second
	}

	// Wrap repositories with timeout middleware to ensure all database operations have bounded execution time
	urlRepo = repository.NewURLRepositoryWithTimeout(urlRepo, dbTimeout)
	clickRepo = repository.NewClickRepositoryWithTimeout(clickRepo, dbTimeout)

	// Initialize URL generator
	generator, err := url.NewGenerator(urlRepo, url.DefaultGeneratorConfig())
	if err != nil {
		return nil, fmt.Errorf("failed to create URL generator: %w", err)
	}

	// Initialize use cases
	createUseCase := application.NewCreateURLUseCase(generator, s.config.BaseURL)
	listUseCase := application.NewListURLsUseCase(urlRepo, clickRepo)
	deleteUseCase := application.NewDeleteURLUseCase(urlRepo)
	getAnalyticsUseCase := application.NewGetAnalyticsUseCase(urlRepo, clickRepo)
	s.redirectUseCase = application.NewRedirectURLUseCaseWithOptions(urlRepo, clickRepo, application.RedirectURLOptions{
		MaxWorkers: s.config.RedirectClickWorkers,
		QueueSize:  s.config.RedirectClickQueueSize,
		Logger:     &s.logger,
		Metrics:    s.metrics,
		StatusRepo: urlStatusRepo,
	})

	s.urlStatusChecker = application.NewURLStatusChecker(urlStatusRepo, application.URLStatusCheckerConfig{
		Enabled:                s.config.URLStatusCheckerEnabled,
		PollInterval:           s.config.URLStatusCheckerPollInterval,
		AliveRecheckInterval:   s.config.URLStatusCheckerAliveRecheckInterval,
		GoneRecheckInterval:    s.config.URLStatusCheckerGoneRecheckInterval,
		BatchSize:              s.config.URLStatusCheckerBatchSize,
		Concurrency:            s.config.URLStatusCheckerConcurrency,
		ArchiveLookupEnabled:   s.config.URLStatusCheckerArchiveLookupEnabled,
		ArchiveRecheckInterval: s.config.URLStatusCheckerArchiveRecheckInterval,
	}, s.logger)

	// Initialize handlers
	urlHandler := handlers.NewURLHandler(createUseCase, listUseCase, deleteUseCase)
	analyticsHandler := handlers.NewAnalyticsHandler(getAnalyticsUseCase)
	redirectHandler := handlers.NewRedirectHandler(s.redirectUseCase)
	pageHandler := handlers.NewPageHandler(createUseCase, listUseCase, s.config.ActiveAuthTokens(), s.sessionStore, s.config.SecureCookies)

	return &routeHandlers{
		urlHandler:       urlHandler,
		analyticsHandler: analyticsHandler,
		redirectHandler:  redirectHandler,
		pageHandler:      pageHandler,
	}, nil
}

func (s *Server) setupPageRoutes(pageHandler *handlers.PageHandler) {
	// HTML page routes
	s.router.Get("/", pageHandler.Home)
	s.router.HandleFunc("/create", pageHandler.CreatePage)
	s.router.HandleFunc("/login", pageHandler.Login)
	s.router.Get("/logout", pageHandler.Logout)

	// Protected dashboard route (requires session)
	s.router.With(middleware.RequireSession(s.sessionStore, "/login")).Get("/dashboard", pageHandler.Dashboard)
}

func (s *Server) setupRedirectRoutes(redirectHandler *handlers.RedirectHandler, redirectRateLimiter *middleware.RateLimiterMiddleware) {
	s.router.With(redirectRateLimiter.Middleware).Get("/{shortCode}", redirectHandler.Redirect)
}

func (s *Server) setupAPIRoutes(urlHandler *handlers.URLHandler, analyticsHandler *handlers.AnalyticsHandler, apiRateLimiter *middleware.RateLimiterMiddleware) {
	// API routes with authentication
	s.router.Route("/api", func(r chi.Router) {
		r.Use(apiRateLimiter.Middleware)

		r.Route("/urls", func(r chi.Router) {
			// Apply auth middleware to all URL endpoints
			// Support both Bearer token auth (for API) and session auth (for dashboard)
			r.Use(middleware.SessionOrBearerAuth(s.config.ActiveAuthTokens()))

			r.Post("/", urlHandler.Create)
			r.Get("/", urlHandler.List)
			r.Delete("/{shortCode}", urlHandler.Delete)
			r.Get("/{shortCode}/analytics", analyticsHandler.GetAnalytics)
		})
	})
}

// healthCheckHandler returns a simple health check response
func (s *Server) healthCheckHandler(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)
	_, _ = w.Write([]byte(`{"status":"ok"}`))
}

// readyCheckHandler returns readiness status and validates external dependencies.
func (s *Server) readyCheckHandler(w http.ResponseWriter, r *http.Request) {
	timeout := s.config.DBTimeout
	if timeout <= 0 {
		timeout = 5 * time.Second
	}

	ctx, cancel := context.WithTimeout(r.Context(), timeout)
	defer cancel()

	w.Header().Set("Content-Type", "application/json")

	if s.db == nil {
		w.WriteHeader(http.StatusServiceUnavailable)
		_, _ = w.Write([]byte(`{"status":"unavailable"}`))
		return
	}

	if err := s.db.PingContext(ctx); err != nil {
		w.WriteHeader(http.StatusServiceUnavailable)
		_, _ = w.Write([]byte(`{"status":"unavailable"}`))
		return
	}

	w.WriteHeader(http.StatusOK)
	_, _ = w.Write([]byte(`{"status":"ready"}`))
}

// Start starts the HTTP server
func (s *Server) Start() error {
	if s.urlStatusChecker != nil {
		s.urlStatusChecker.Start(s.bgCtx)
	}

	s.logger.Info().Str("addr", s.httpServer.Addr).Msg("starting HTTP server")
	return s.httpServer.ListenAndServe()
}

// Shutdown gracefully shuts down the server
func (s *Server) Shutdown(ctx context.Context) error {
	s.logger.Info().Msg("shutting down HTTP server")

	if s.bgCancel != nil {
		s.bgCancel()
	}
	if s.urlStatusChecker != nil {
		s.urlStatusChecker.Shutdown()
	}

	// Shutdown session store cleanup goroutine
	if s.sessionStore != nil {
		s.sessionStore.Shutdown()
	}

	// Shutdown redirect use case workers
	if s.redirectUseCase != nil {
		s.redirectUseCase.Shutdown()
	}

	for _, limiter := range s.rateLimiters {
		limiter.Shutdown()
	}

	if err := s.httpServer.Shutdown(ctx); err != nil {
		return fmt.Errorf("failed to shutdown server: %w", err)
	}

	s.logger.Info().Msg("HTTP server stopped")
	return nil
}

// Router returns the chi router for testing purposes
func (s *Server) Router() *chi.Mux {
	return s.router
}

// Metrics returns the Prometheus metrics for the server
func (s *Server) Metrics() *metrics.Metrics {
	return s.metrics
}